scale up and refactor plotting script
