+++ asf_heat_pump_suitability/analysis/hn_zones/config/hnz_config.py
@@ -21,6 +21,9 @@ from asf_heat_pump_suitability import PROJECT_DIR
 # Output Directory
 OUTPUT_DIR = os.path.join(PROJECT_DIR, "outputs/hn_zones/output_data/")
 
+OUTPUT_PLOTS_DIR = os.path.join(PROJECT_DIR, "outputs/hn_zones/output_plots_all/")
+
+
 # LOCAL AUTHORITY DICTIONARIES
 LOCAL_AUTHORITIES = {
     "Birmingham": "heat-network-zone-map-Birmingham.gpkg",
@@ -71,18 +74,42 @@ LSOA_SHP_PATH_LOCAL = os.path.join(
 )
 LSOA_SHP_PATH_S3 = "s3://asf-heat-pump-suitability/source_data/Lower_layer_Super_Output_Areas_2021_EW_BFE_V9_-9107090204806789093/LSOA_2021_EW_BFE_V9.shp"
 
+# NESTA_HPS_PARQUET_LOCAL = os.path.join(
+#    PROJECT_DIR,
+#    "asf_heat_pump_suitability/analysis/hn_zones/input_data/nesta_heat_network_suitability/20240925_2023_Q4_EPC_heat_pump_suitability_per_lsoa.parquet",
+# )
+# NESTA_HPS_PARQUET_LOCAL = os.path.join(
+#    PROJECT_DIR,
+#    "asf_heat_pump_suitability/analysis/hn_zones/input_data/nesta_heat_network_suitability/20250206_2023_Q4_heat_pump_suitability_per_lsoa.parquet",
+# )
 NESTA_HPS_PARQUET_LOCAL = os.path.join(
     PROJECT_DIR,
-    "asf_heat_pump_suitability/analysis/hn_zones/input_data/nesta_heat_network_suitability/20240925_2023_Q4_EPC_heat_pump_suitability_per_lsoa.parquet",
+    "asf_heat_pump_suitability/analysis/hn_zones/input_data/nesta_heat_network_suitability/20250305_2023_Q4_heat_pump_suitability_per_lsoa.parquet",
 )
-NESTA_HPS_PARQUET_S3 = "s3://nesta-open-data/asf_heat_pump_suitability/2023Q4/20240925_2023_Q4_EPC_heat_pump_suitability_per_lsoa.parquet"
+
+# NESTA_HPS_PARQUET_S3 = "s3://nesta-open-data/asf_heat_pump_suitability/2023Q4/20240925_2023_Q4_EPC_heat_pump_suitability_per_lsoa.parquet"
+# NESTA_HPS_PARQUET_S3 = "s3://asf-heat-pump-suitability/outputs/2023Q4/suitability/20250206_2023_Q4_heat_pump_suitability_per_lsoa.parquet"
+NESTA_HPS_PARQUET_S3 = "s3://asf-heat-pump-suitability/outputs/2023Q4/suitability/20250305_2023_Q4_heat_pump_suitability_per_lsoa.parquet"
+
+NESTA_HPS_PARQUET_PATHS = {
+    "local": NESTA_HPS_PARQUET_LOCAL,
+    "s3": NESTA_HPS_PARQUET_S3,
+}
 
 S3_BUCKET = "asf-heat-pump-suitability"
 S3_KEY_DIR = "evaluation/desnz_hn_zone_scores/"
 
+# Target CRS (ensure LSOA shapefiles and GPKGs are converted if they differ)
+TARGET_CRS = "EPSG:27700"
+
 
 # Default DESNZ Pilot Fraction Threshold
 DEFAULT_THRESHOLD = 0.0  # This can be modified here instead of in multiple scripts
+
+ABSOLUTE_ERROR_THRESHOLD_ABSENT = 0.0  # Threshold for 'abs_error' when DESNZ HN zone is absent (if change from 0, this no means no longer absent)
+ABSOLUTE_ERROR_THRESHOLD_PRESENT = (
+    0.000001  # Threshold for 'abs_error' when DESNZ HN zone is present
+)
 
 # List of Thresholds for Score Calculation
 THRESHOLDS = [round(i * 0.05, 2) for i in range(0, 20)]  # [0.0, 0.05, ..., 0.95]
